Added ReviewModal and DetailPanel. Seems to be working
