+++ backend/frontend/src/components/Loginform.jsx
@@ -6,77 +6,75 @@ import { useToast } from '@chakra-ui/react'
 
 function Loginform() {
 
-  const [formDetails, setFormDetails] = useState({ email: "", password: "" });
-  const [error, setError] = useState("");
-  const navigate = useNavigate();
-  const toast = useToast()
+	const [formDetails, setFormDetails] = useState({ email: "", password: "" });
+	const [error, setError] = useState("");
+	const navigate = useNavigate();
+	const toast = useToast()
 
-  const loginHandler = async (e) => {
+	const loginHandler = async (e) => {
 
-    e.preventDefault();
-    // console.log("formDetails", formDetails)
-    const apiService = new ApiService();
+		e.preventDefault();
+		const apiService = new ApiService();
 
-    try {
+		try {
 
-      const loginResponse = await apiService.login(formDetails);
-      // console.log("loginResponse", loginResponse)
-      let authenticatedUserObj = {
-        'token': loginResponse['token'],
-        'id': loginResponse['id'],
-        'username': loginResponse['username']
-      }
-      localStorage.setItem('authenticatedUser', JSON.stringify(authenticatedUserObj));
-      window.dispatchEvent(new Event("storage"));
-      toast({
-        title: 'Login Successful',
-        status: 'success',
-        duration: 4000,
-        isClosable: true,
-        position: 'bottom-right',
-        variant: 'left-accent'
-      })
+			const loginResponse = await apiService.login(formDetails);
+			let authenticatedUserObj = {
+				'token': loginResponse['token'],
+				'id': loginResponse['id'],
+				'username': loginResponse['username']
+			}
+			localStorage.setItem('authenticatedUser', JSON.stringify(authenticatedUserObj));
+			window.dispatchEvent(new Event("storage"));
+			toast({
+				title: 'Login Successful',
+				status: 'success',
+				duration: 4000,
+				isClosable: true,
+				position: 'bottom-right',
+				variant: 'left-accent'
+			})
 
-      navigate("/home");
+			navigate("/home");
 
-    } catch (error) {
+		} catch (error) {
 
-      console.error(error['non_field_errors'])
-      setError(error['non_field_errors'][0])
+			console.error(error['non_field_errors'])
+			setError(error['non_field_errors'][0])
 
-    }
-  }
+		}
+	}
 
-  return (
-    <div className="login">
-      <div className="form-container">
-        <form onSubmit={loginHandler}>
-          <div className="form-inner">
+	return (
+		<div className="login">
+			<div className="form-container">
+				<form onSubmit={loginHandler}>
+					<div className="form-inner">
 
-            <h2>Login Form</h2>
-            <br></br>
+						<h2>Login Form</h2>
+						<br></br>
 
-            {(error !== "") ? (<div className="warning">{error}</div>) : ""}
+						{(error !== "") ? (<div className="warning">{error}</div>) : ""}
 
-            <div className="form-group">
-              <label htmlFor="email">Email: </label>
-              <input type="email" name="email" id="email" onChange={e => setFormDetails({ ...formDetails, email: e.target.value })} value={formDetails.email} />
-            </div>
+						<div className="form-group">
+							<label htmlFor="email">Email: </label>
+							<input type="email" name="email" id="email" onChange={e => setFormDetails({ ...formDetails, email: e.target.value })} value={formDetails.email} />
+						</div>
 
-            <div className="form-group">
-              <label htmlFor="password">Password: </label>
-              <input type="password" name="password" id="password" onChange={e => setFormDetails({ ...formDetails, password: e.target.value })} value={formDetails.password} />
-            </div>
+						<div className="form-group">
+							<label htmlFor="password">Password: </label>
+							<input type="password" name="password" id="password" onChange={e => setFormDetails({ ...formDetails, password: e.target.value })} value={formDetails.password} />
+						</div>
 
-            <input type="submit" value="Login" />
+						<input type="submit" value="Login" />
 
-          </div>
-        </form>
-        <br /><br />
-        <p className="signup-prompt">New user? <Link to='/signup'>Sign up here</Link></p>
-      </div>
-    </div>
-  )
+					</div>
+				</form>
+				<br /><br />
+				<p className="signup-prompt">New user? <Link to='/signup'>Sign up here</Link></p>
+			</div>
+		</div>
+	)
 }
 
 export default Loginform
