+++ backend/frontend/src/components/ReviewList.jsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import ListGroup from 'react-bootstrap/ListGroup';
 import { ApiService } from '../api-service';
-import Button from 'react-bootstrap/Button';
 import { IconButton, Stack, Flex, Spacer } from '@chakra-ui/react';
 import { AiOutlineLike, AiOutlineDislike } from 'react-icons/ai';
 import { BiFlag } from 'react-icons/bi';
@@ -29,7 +28,6 @@ export const ReviewList = (props) => {
     const updateReview = async (updatedReview) => {
         try {
             const updatedReviewResponse = await apiService.updateReview(updatedReview);
-            // console.log("updatedReviewResponse", updatedReviewResponse)
             const reviewData = await apiService.getReview(selectedAmenity, updatedReview.amenity_id);
             setReviews(reviewData);
         }
