+++ backend/frontend/src/styles/App.css
@@ -11,10 +11,9 @@
 .login {
   height: 100vh;
   display: flex;
-  /* align-items: center; */
   margin-top: 100px;
   justify-content: center;
-  background-color :#ffffff;
+  background-color: #ffffff;
 
 }
 
@@ -23,7 +22,8 @@ form {
   position: relative;
 }
 
-input, button {
+input,
+button {
   appearance: none;
   background: none;
   border: none;
